feat(#82): port _editor Flask app to _django subpackage + delete Flask
Port the existing Flask-based `src/scitex_writer/_editor/` to a new Django
subpackage `src/scitex_writer/_django/`, mirroring `figrecipe/_django` so
scitex-cloud's `writer_app` can consume it as a thin wrapper.

Changes:
- New `_django/` tree: apps.py, urls.py, views.py, services.py, settings.py,
  manifest.json, handlers/{core,files,compile,bib,claim}.py, templates/writer/
  editor.html, static/writer/{css,js}
- `ProjectState` dataclass replaces Flask `WriterEditor`; per-project cache w/ TTL
- `api_dispatch` catch-all with HANDLERS registry + method gating, plus
  parameterized `api/claims/<id>[/chain]` fallbacks
- JS updated to read `API_BASE` from `<body data-api-base>` so a single bundle
  serves both local (mounted at `/`) and cloud (mounted at `/writer/`)
- `scitex-writer gui` now launches Django via `_django._server` (runserver)
- `sw.gui` re-exports `_django._server.run`
- pyproject: `django>=4.2` added as core dep; `flask` dropped from `[editor]`
  extra (kept for bwd-compat as empty); new `[desktop]` extra (pywebview)
- Tests: `tests/python/_django/test_endpoints.py` — 14 parity tests covering
  every endpoint the Flask version exposed; full suite: 583 passed
- **Deleted**: `src/scitex_writer/_editor/` (Flask tree + 1,877-LOC Python-string
  templates) — clean port, no dual-maintain, per issue #82

Viewer module (claims metadata, DAG, citation verification) lands in PR2.

Closes #82 (editor port). Follow-up: PR2 viewer module; scitex-cloud#146 to
consume this as a thin wrapper.
