+++ pyproject.toml
@@ -42,6 +42,7 @@ classifiers = [
 dependencies = [
     "fastmcp>=2.0.0",
     "scitex-dev>=0.1.0",
+    "django>=4.2",
 ]
 
 [project.optional-dependencies]
@@ -52,7 +53,10 @@ dev = [
     "pre-commit>=3.5.0",
 ]
 editor = [
-    "flask>=2.0.0",
+    # Django is now a core dep; this group kept for backwards-compat of `[editor]` extra
+]
+desktop = [
+    "pywebview>=4.0.0",
 ]
 docs = [
     "sphinx>=7.0",
@@ -62,8 +66,6 @@ docs = [
     "sphinx-autodoc-typehints>=1.25",
 ]
 all = [
-    # editor
-    "flask>=2.0.0",
     # dev
     "pytest>=7.0.0",
     "pytest-cov>=4.0.0",
@@ -102,7 +104,7 @@ addopts = "-v --tb=short"
 
 [tool.coverage.run]
 omit = [
-    "src/scitex_writer/_editor/*",
+    "src/scitex_writer/_django/*",
     "src/scitex_writer/_mcp/handlers/*",
     "src/scitex_writer/_mcp/tools/*",
 ]
