+++ src/scitex_writer/_django/settings.py
@@ -1,0 +1,49 @@
+#!/usr/bin/env python3
+# -*- coding: utf-8 -*-
+"""Minimal standalone Django settings for `scitex-writer gui`.
+
+Enough to run `django.core.management runserver`; cloud deployments
+do NOT import this file — they use their own settings and just mount
+`scitex_writer._django.urls` under a prefix (see scitex-cloud
+`writer_app/urls/__init__.py`).
+"""
+
+from __future__ import annotations
+
+import os
+import secrets
+from pathlib import Path
+
+BASE_DIR = Path(__file__).resolve().parent
+
+SECRET_KEY = os.environ.get("WRITER_DJANGO_SECRET", secrets.token_urlsafe(32))
+DEBUG = True
+ALLOWED_HOSTS = ["*"]
+
+INSTALLED_APPS = [
+    "django.contrib.staticfiles",
+    "scitex_writer._django.apps.WriterEditorConfig",
+]
+
+MIDDLEWARE = [
+    "django.middleware.common.CommonMiddleware",
+]
+
+ROOT_URLCONF = "scitex_writer._django._standalone_urls"
+
+TEMPLATES = [
+    {
+        "BACKEND": "django.template.backends.django.DjangoTemplates",
+        "DIRS": [],
+        "APP_DIRS": True,
+        "OPTIONS": {
+            "context_processors": [
+                "django.template.context_processors.request",
+            ],
+        },
+    },
+]
+
+STATIC_URL = "/static/"
+DEFAULT_AUTO_FIELD = "django.db.models.BigAutoField"
+USE_TZ = True
